chore: move value_type.rs to common module
