+++ src/common/mod.rs
@@ -2,5 +2,6 @@ mod student_info;
 mod user;
 mod form_schema;
 mod expandable_info;
+pub mod value_type;
 
-pub use self::{student_info::*, user::*, form_schema::*, expandable_info::*};
+pub use self::{expandable_info::*, form_schema::*, student_info::*, user::*};
