+++ src/components/multi_entry/mod.rs
@@ -1,11 +1,10 @@
 pub mod multi_entry_data;
-pub mod value_type;
 pub mod multi_entry;
 mod multi_entry_member;
 
+pub use crate::common::value_type::*;
 pub use self::{
-    value_type::*,
-    multi_entry_data::*,
     multi_entry::*,
+    multi_entry_data::*,
     multi_entry_member::*,
 };
